Support for deploy directory to be consumed by chectl.
The "meat" is in the Makefile and the deploy/generate-deployment.sh script,
and the deploy/templates kustomize declarations.
the deploy/deployment directory is generated but needs to be in the source
control so that chectl can directly consume it.

Also adding a github action on PRs to make sure all the generated files are
up-to-date before merging.